Fix half-closed TCP may hang on forever
When one of either upstream proxy or local client half close the TCP
connection, these two TCPs will enter (CLOSE_WAIT, FIN_WAIT_2) state.

Normally, this state will exit once the other side close its connection
with a FIN sent to us. But if it doesn't, the pairs of connection will
hang on forever.

To prevent accumulation of such dead connections, a timer is added.
Half-closed connections will be forcibly closed once there is no traffic
after 60 seconds (same as the default value of Linux's tcp_fin_timeout).

+++ src/main.rs
@@ -101,7 +101,7 @@ async fn main() {
         .value_of("n-parallel")
         .parse()
         .expect("not a valid number")
-        .unwrap_or(0 as usize);
+        .unwrap_or(0_usize);
     let cong_local = args.value_of("cong-local");
     let allow_direct = args.is_present("allow-direct");
     let graphite = args
